test-maven-plugin/src/main/java/com/uitnetwork/TestMojo.java
package com.uitnetwork;


import org.apache.maven.plugin.AbstractMojo;
import org.apache.maven.plugin.MojoExecutionException;

import org.apache.maven.plugins.annotations.LifecyclePhase;
import org.apache.maven.plugins.annotations.Mojo;
import org.apache.maven.plugins.annotations.Parameter;

import java.io.File;
import java.io.FileWriter;
import java.io.IOException;

@Mojo(name = "test", defaultPhase = LifecyclePhase.COMPILE)
public class TestMojo extends AbstractMojo {
    /**
     * Location of the file.
     */
    @Parameter(defaultValue = "${project.build.directory}", property = "outputDir", required = true)
    private File outputDirectory;

    public void execute()
            throws MojoExecutionException {
        getLog().info("This is a test plugin to do something with dir: " + outputDirectory);
    }
}
