test-maven-plugin/src/test/java/com/uitnetwork/TestMojoTest.java
package com.uitnetwork;


import org.apache.maven.plugin.testing.MojoRule;
import org.apache.maven.plugin.testing.WithoutMojo;

import org.junit.Rule;

import static org.junit.Assert.*;

import org.junit.Test;

import java.io.File;

public class TestMojoTest {
    @Rule
    public MojoRule rule = new MojoRule() {
        @Override
        protected void before() throws Throwable {
        }

        @Override
        protected void after() {
        }
    };

    /**
     * @throws Exception if any
     */
    @Test
    public void testSomething()
            throws Exception {
        File pom = new File("target/test-classes/project-to-test/");
        assertNotNull(pom);
        assertTrue(pom.exists());

        TestMojo testMojo = (TestMojo) rule.lookupConfiguredMojo(pom, "test");
        assertNotNull(testMojo);
        testMojo.execute();

        File outputDirectory = (File) rule.getVariableValueFromObject(testMojo, "outputDirectory");
        assertNotNull(outputDirectory);
    }

    /**
     * Do not need the MojoRule.
     */
    @WithoutMojo
    @Test
    public void testSomethingWhichDoesNotNeedTheMojoAndProbablyShouldBeExtractedIntoANewClassOfItsOwn() {
        assertTrue(true);
    }

}

